refactor: Extract reusable code from CLI commands (`mfa configure`, `load-variables`)
